backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
langchain==0.3.14
langchain-openai==0.2.14
langgraph==0.2.59
chromadb==0.5.23
pypdf==5.1.0
langchain-community==0.3.14
